+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -499,6 +499,14 @@ public class EntityPanel extends JPanel {
             .anyMatch(detailPanel -> detailPanel.entityModel.entityType().equals(entityType));
   }
 
+  /**
+   * Selects the given detail panel if it is available
+   * @param detailPanel the detail panel to select
+   */
+  public final void selectDetailPanel(EntityPanel detailPanel) {
+    detailLayout().select(requireNonNull(detailPanel));
+  }
+
   @Override
   public final String toString() {
     return getClass().getSimpleName() + ": " + caption.get();
@@ -906,6 +914,14 @@ public class EntityPanel extends JPanel {
     }
   }
 
+  /**
+   * @return the detail layout used by this panel
+   * @param <T> the detail layout type
+   */
+  protected final <T extends DetailLayout> T detailLayout() {
+    return (T) configuration.detailLayout;
+  }
+
   private JPanel createEditControlPanel() {
     if (editPanel == null) {
       return null;
@@ -939,10 +955,6 @@ public class EntityPanel extends JPanel {
 
   final void setNextSiblingPanel(EntityPanel nextSiblingPanel) {
     this.nextSiblingPanel = requireNonNull(nextSiblingPanel);
-  }
-
-  final <T extends DetailLayout> T detailLayout() {
-    return (T) configuration.detailLayout;
   }
 
   static void addEntityPanelAndLinkSiblings(EntityPanel detailPanel, List<EntityPanel> entityPanels) {